detector_model_scripts/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
ca-certificates=2024.7.2=haa95532_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.114.2=pypi_0
filelock=3.16.0=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.9.0=pypi_0
h11=0.14.0=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
importlib-resources=6.4.5=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
lmdb=1.5.1=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
numpy=2.0.2=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.0.15=h827c3e9_0
packaging=24.1=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py39haa95532_0
pydantic=2.9.1=pypi_0
pydantic-core=2.23.3=pypi_0
pyparsing=3.1.4=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
setuptools=72.1.0=py39haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.38.5=pypi_0
svgpathtools=1.6.1=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.13.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
torch=2.4.1=pypi_0
torchvision=0.19.1=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
uvicorn=0.30.6=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py39haa95532_0
zipp=3.20.2=pypi_0
